Inherited Address, Email and Phone from Contact class.
